Assignment2/requirements.txt
google-api-python-client==2.101.0
google-cloud-storage==2.11.0
numpy==1.24.1
beautifulsoup4==4.11.1